Move RDBJobEventSearch to elastic-job-lite-console module
